+++ src/screens/SignUp/constants.ts
@@ -1,1 +1,6 @@
-export const FIELDS = ['firstName', 'lastName', 'email', 'password', 'passwordConfirmation'];
+export interface IFormInput {
+  email: string;
+  firstName: string;
+  lastName: string;
+  password: string;
+}
